requirements.txt
pyzabbix==1.0.0
flake8==3.9.2
flake8-docstrings==1.6.0
python-dotenv==0.19.0
requests==2.26.0